requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scikit-learn>=1.2.0
Pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
Flask>=2.3.0
werkzeug>=2.3.0